Remove a subtest which only passes by accident.
The content of the file around address 0x112f looks as follows:

```asm
           frame_dummy:
401120         jmp        register_tm_clones
401122         align      16
           main:
401130         push       rbp
401131         mov        rbp, rsp
401134         sub        rsp, 0x10
401138         mov        dword [rbp+var_4], 0x0
40113f         mov        dword [rbp+var_8], 0x0
401146         lea        rdi, qword [rbp+var_8]
40114a         call       f
40114f         mov        eax, dword [rbp+var_8]
401152         add        rsp, 0x10
401156         pop        rbp
401157         ret
401158         align      32
```

The reason why we currently don't find a symbol for address 0x112f
is because we ignore the `frame_dummy` symbol entirely.

If we change the symbol filter so that `frame_dummy` is no longer
ignored, then we will consider the alignment gap to be part of
`frame_dummy`, because the `frame_dummy` symbol does not define a size.

And that's ok. I'd rather have too many symbols than too few. So I'm
removing this test.